# myblog/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gin+gorm+jwt+cors+viper等等",
        "title": "myblog gin+gorm",
        "contact": {
            "name": "go-swagger帮助文档",
            "url": "https://github.com/swaggo/swag/blob/master/README_zh-CN.md"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/article/add": {
            "post": {
                "description": "后台添加一个文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "后台添加文章",
                "parameters": [
                    {
                        "description": "添加文章",
                        "name": "article",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.Article"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    }
                }
            }
        },
        "/api/article/delete/{id}": {
            "delete": {
                "description": "后台删除一个文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "后台删除文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "要删除文章的id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    }
                }
            }
        },
        "/api/article/edit/{id}": {
            "put": {
                "description": "后台编辑一个动态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "后台更新文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "要编辑分类的id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "编辑分类",
                        "name": "category",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.Category"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    }
                }
            }
        },
        "/api/article/list": {
            "get": {
                "description": "查看文章列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "查看文章列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "title",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    }
                }
            }
        },
        "/api/article/search": {
            "get": {
                "description": "根据标题检索文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "检索文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文章标题",
                        "name": "title",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    }
                }
            }
        },
        "/api/article/{id}": {
            "get": {
                "description": "前台查看文章内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "前台查看文章内容",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "要查看的id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    }
                }
            }
        },
        "/api/articlecate/{name}": {
            "get": {
                "description": "查看分类文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "前台查看分类文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分类名字",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分页尺寸",
                        "name": "pagesize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pagenum",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Article": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "integer"
                },
                "content": {
                    "type": "string",
                    "example": "content"
                },
                "desc": {
                    "type": "string",
                    "example": "desc"
                },
                "img": {
                    "type": "string",
                    "example": "img"
                },
                "read_ount": {
                    "type": "integer"
                },
                "title": {
                    "type": "string",
                    "example": "title"
                }
            }
        },
...
        "model.Category": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "name"
                }
            }
        },
...
        "model.Reponse": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}